+++ repository/src/main/java/org/grnet/cat/repositories/registry/PrincipleRepository.java
@@ -1,16 +1,17 @@
-package org.grnet.cat.repositories;
+package org.grnet.cat.repositories.registry;
 
-import io.quarkus.hibernate.orm.panache.PanacheRepository;
 import jakarta.enterprise.context.ApplicationScoped;
 import org.grnet.cat.entities.Page;
 import org.grnet.cat.entities.PageQuery;
 import org.grnet.cat.entities.PageQueryImpl;
-import org.grnet.cat.entities.Principle;
+import org.grnet.cat.entities.registry.Principle;
+import org.grnet.cat.repositories.Repository;
+
 import java.util.HashMap;
 
 import java.util.StringJoiner;
 @ApplicationScoped
-public class PrincipleRepository implements PanacheRepository<Principle> {
+public class PrincipleRepository implements Repository<Principle, String> {
 
     /**
      * Fetches guidance items by page.
@@ -20,12 +21,13 @@ public class PrincipleRepository implements PanacheRepository<Principle> {
      * @return A PageQuery of guidance items.
      */
     public PageQuery<Principle> fetchPrincipleByPage(int page, int size) {
+
         var joiner = new StringJoiner(" ");
         joiner.add("from Principle p");
 
         var params = new HashMap<String, Object>();
 
-        joiner.add("order by p.createdOn desc");
+        joiner.add("order by p.populatedBy desc");
 
         var panacheQuery = find(joiner.toString(), params).page(page, size);
 
@@ -46,6 +48,5 @@ public class PrincipleRepository implements PanacheRepository<Principle> {
                 .getSingleResult();
         return count > 0;
     }
-
 }